test(ui): the two allocation gates stop racing each other
The widget tree's allocation gates shared a file, and a global
allocator is process-wide while cargo runs one file's tests
concurrently. So the two raced: one opened its measured window while
the other was still freeing its fixtures, and the loser reported a
delta it had not caused.

It surfaced as a lane failing with one allocation against twelve
deallocations — more frees than allocations, which is the signature of
counting someone else's teardown, and which no amount of staring at the
test that reported it would explain.

The fix is the rule the sibling crates already state in their own
source: one counting test per file. Separate files are separate
binaries and separate processes, so the counters mean what they say.
Both tests keep their full shape and their premises; only the file
boundary moves.

That this had to be learned from a red lane rather than read from the
neighbours is the part worth recording, so both files now say why they
are separate rather than leaving the next author to discover it the
same way.
